\name{phase1.1.qc.test}
\alias{phase1.1.qc.test}
\alias{phase1.1.qc.test}
\docType{package}
\title{
Quality control for phase1.1 data
}
\description{
The package \pkg{phase1.1.qc.test} consists of a collection of functions for detecting potential quality control issues.
}
\details{
\tabular{ll}{
Package: \tab phase1.1.qc.test\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2020-8-19\cr
License: GPL>=2 \cr
}
}

\author{
Chuan Hong \cr
Maintainer: Chuan Hong <chuan_hong@hms.harvard.edu>
}

\keyword{package}

